requirements.txt
requests==2.31.0
python-dotenv==0.20.0
Flask==2.2.5
gunicorn==21.2.0
gevent
urllib3==1.26.12
flask-openapi3==2.5.4
Flask-Limiter==3.5.0
bittensor